STYLE : DetailPage

+++ src/components/Board/BoardDetail.js
@@ -39,23 +39,25 @@ const BoardDetail = () => {
   const findItem = post.find((it) => {
     return parseInt(it.no) == parseInt(data.no);
   });
-  console.log(findItem);
+  // console.log(findItem); 확인용
   //삭제 기능
   const removePosts = async () => {
-    if (findItem.userId == userInfo[1]) {
-      try {
-        const response = await axios.delete(
-          `http://localhost:8080/api-board/delete?board_no=${findItem.board_no}`,
-        );
-        console.log(response);
-        alert(`글을 삭제합니다.`);
-        navigate(`/${board}`, { replace: true });
-      } catch (error) {
-        console.log(error);
+    if (window.confirm('게시글을 삭제하시겠습니까?')) {
+      if (findItem.userId == userInfo[1]) {
+        try {
+          const response = await axios.delete(
+            `http://localhost:8080/api-board/delete?board_no=${findItem.board_no}`,
+          );
+          // console.log(response); api 확인용
+          navigate(`/${board}`, { replace: true });
+        } catch (error) {
+          console.log(error);
+        }
+        alert('글을 삭제합니다');
+        getData();
+      } else {
+        alert('권한이 없습니다');
       }
-      getData();
-    } else {
-      alert('권한이 없습니다');
     }
   };
   return (
@@ -69,25 +71,33 @@ const BoardDetail = () => {
             <span>작성 시간 {findItem.create_dt}</span>
           </div>
           {findItem.userId == userInfo[1] ? (
-            <div className={style.btnWrapper}>
-              <button onClick={removePosts}>삭제</button>
+            <div className={style.titlebtnWrapper}>
               <button
                 onClick={() => {
-                  navigate(`/${board}/${no}/edit`);
+                  if (window.confirm('게시글을 수정하시겠습니까')) {
+                    navigate(`/${board}/${no}/edit`);
+                  }
                 }}
               >
                 수정
               </button>
+              <button onClick={removePosts}>삭제</button>
             </div>
           ) : null}
           <div className={style.boardText}>
             <div dangerouslySetInnerHTML={{ __html: findItem.content }} />
             {/* 스타일을 읽을 라면 이거 써야된다고 함 */}
           </div>
-          <button onClick={() => navigate(`/${board}/${++no}`)}>다음글</button>
-          <button onClick={() => navigate(`/${board}/${no - 1}`)}>
-            이전글
-          </button>
+          <div className={style.btnWrapper}>
+            <div>
+              <button onClick={() => navigate(`/${board}/${++no}`)}>
+                다음글
+              </button>
+              <button onClick={() => navigate(`/${board}/${no - 1}`)}>
+                이전글
+              </button>
+            </div>
+          </div>
         </div>
       </Board>
       <CommentsWrapper>
